+++ src/apis/auth.ts
@@ -1,1 +1,61 @@
 /* 로그인 관련 API 요청 */
+
+import axios from 'axios';
+
+const getGoogleProcess = async (code: string | null): Promise<boolean | undefined> => {
+  if (code === null || !code) {
+    return false;
+  }
+
+  try {
+    const res = await axios.get(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/login/google/process`);
+    const data = res.data;
+    console.log('구글 로그인 데이터:', data);
+    localStorage.setItem('accessToken', data.access_token);
+    localStorage.setItem('refreshToken', data.refresh_token);
+    return true;
+  } catch (err) {
+    alert(`구글 로그인 에러 발생 : ${err}`);
+    return false;
+  }
+};
+
+const getNaverProcess = async (code: string | null): Promise<boolean | undefined> => {
+  if (code === null || !code) {
+    return false;
+  }
+  try {
+    const res = await axios.get(
+      `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/login/naver/process?code=${code}`,
+    );
+    const data = await res.data;
+    console.log('네이버 로그인 데이터:', data);
+    localStorage.setItem('accessToken', data.access_token);
+    localStorage.setItem('refreshToken', data.refresh_token);
+    return true;
+  } catch (err) {
+    alert(`네이버 로그인 에러 발생 : ${err}`);
+    return false;
+  }
+};
+
+const getKakaoProcess = async (code: string | null): Promise<boolean | undefined> => {
+  if (code === null || !code) {
+    return false;
+  }
+  try {
+    const res = await axios.get(
+      `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/login/kakao/process?code=${code}`,
+    );
+    const data = await res.data;
+    console.log('카카오 로그인 데이터:', data);
+    localStorage.setItem('accessToken', data.access_token);
+    localStorage.setItem('refreshToken', data.refresh_token);
+    return true;
+  } catch (err) {
+    alert(`카카오 로그인 에러 발생 : ${err}`);
+    return false;
+  }
+};
+
+export { getNaverProcess, getKakaoProcess, getGoogleProcess };
